Add coroutines and serialization dependencies, add json files, add 2 model class: Actor.kt and Genre.kt, some code refactor
