+++ app/src/main/java/com/johnsondev/doboshacademyapp/model/Actor.kt
@@ -1,0 +1,9 @@
+package com.johnsondev.doboshacademyapp.model
+
+import java.io.Serializable
+
+data class Actor(
+    val id: Int,
+    val name: String,
+    val imageUrl: String,
+) : Serializable
