+++ app/src/main/java/com/johnsondev/doboshacademyapp/model/Genre.kt
@@ -1,0 +1,4 @@
+package com.johnsondev.doboshacademyapp.model
+
+
+data class Genre(val id: Int, val name: String)
